pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hosted-agents-saas-platform"
version = "0.1.0"
description = "SMB multitenant control plane and execution gateway for Foundry agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "azure-ai-projects>=1.0.0b10,<2.0.0",
  "azure-identity>=1.17.0",
  "azure-keyvault-secrets>=4.8.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.0",
  "psycopg[binary]>=3.2.0",
  "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
